Fix `sample` bug
R is truly and utterly evil. It has no scalar, everything is a vector.
But we often really need a scalar, so we pretend. In this particular
case, the function `sample` treats `x` as a set from which to sample in
all cases except when `x` is a numeric vector of length 1, in which case
it is treated as the upper bound of a discrete distribution on 1 to `x`.

I looped over a list of integer vectors (all holding taxon ids),
sampling a random element from each. However, when one of the vectors
happened to be of length 1, then I got back a random id. Holy shit! I
solve this by treating the ids as characters.

Gah! I burned two hours tracking down this bug. If I hadn't been
careful, I could easily have missed it altogether until the analysis far
downstream. I really need to stop using this language. But damn does it
have a fine package system. And data munging and visualization ...