Set travis to use NCBI-BLAST v2.6.0
Currently travis in in Ubuntu trusty and apt-get installs v2.2. The
BLAST test is fails, downloading fewer rows of data than I would expect.
I am not sure this is due to a difference in BLAST versions. If it is,
it is a pretty big difference, perhaps a different threshold for keeing
hits? I should look into the changelog ...