Move taxonomy.R to ncbi-taxonomy.R, add ncbi_tree
All the code that previously was in `taxonomy.R` are really specific to
the NCBI common tree handling, so I have renamed the file to represent
this.

I have also added an `ncbi_tree` function. This function takes a list of
taxon IDs and retrieves an NCBI common tree from them. `taxize` roughly
accomplished this, but does not add labels to the internal nodes. I may
be able to get by without such labels, and will do so for now. If
needed, I can find these labels with some `mrca` function.