Converted uniprot_cousins to phylo tree output
Originally I did some filtering in this function.  But now I think that
is unnecessary.  It is quite fast, actually to just download all the
UniProt taxon IDs and build a tree (only takes around 30s).  Also it is
more modular to move the filtering into a different function.  Anyway,
since this function now always gives the same tree (assuming the UniProt
data hasn't changed), maybe it would be better to just store the result.

I ran into a few tricky problems.  First of all, I can't build trees
where a taxon is both a tip and a node.  `ape` will die if I try.  This
problem occurs when both a species and subspecies have represented
proteomes. My solution was to filter out any subspecies where a
species-terminating lineage is present.  I also ran into issues with one
ID that seems to be missing in NCBI.  About this ID, I need a warning
for the case.