In Strata, use one label for the focal species
Previously, I stored the focal species and focal id in the Strata
object. This was bad for two reasons:

 1. Not all taxa have NCBI taxonomy IDs. For example, simulated data,
    unranked species, sub-species, specific individuals, specific genome
    assemblies/annotations, etc.

 2. Any function needing to access the focal species, needed to be told
    (or infer) whether the tips were in terms of ids or names.

Now the the Strata object has the slot 'focal_species' which is
guaranteed to always match the focal species name in the tree. I added a
function `is_valid_strata` that asserts this (among other things).

In this commit, I also adapt the 'make_tree_relative_to' function to
Strata objects with, 'sort_strata', which always sorts the tree relative
to the focal species.