In strata_blast, automatically find focal sequence
Previously strata_blast took a query, however, the function (and
downstream code) assumes the blast is of the focal species against the
related species, so it is reasonable to just directly use the code
listed in the `Stata@data$faa[[focal_species]]` field. This is now done.
One less thing to worry about.