Add `replace_branch` function
The function is nitroglycerin. If you look at it funny, it will blow up
in your face. More precisely, it internally relies on `add_taxa`, which
is wicked little thing that claims to assume all nodes in the input tree
are NCBI taxonomy IDs. However, the ad hoc polymorphism of taxizedb
actually allows names to be used in place of IDs. So if you send in a
named tree, you get back an ID tree.

`replace_branch` works fine if all trees in involved are taxid trees.
But if any of them are named, everything goes to hell.

My desire for IDs and names to coexist in harmony is proving to be my
undoing.

The other demon in my code is `ape`. Every operation in `ape` requires
tinkering with the internal implementation of the tree. There is no
concept of recusion. No recursion in a tree data structure. Madness.