Add analysis of FWER to case studies
Just some rudimentary exploration at this point. The goal is to have
somewhat more interpretable results and thresholds. It does not appear
to make a huge difference, but at least it is better than the arbitrary
E-value cutoff.

I'll need to teak it a bit more. The goal is to have a probability
estimate for the classification of gene into each phylostratum.

Currently I roughly have the probability that each gene is present in
each phylostratum. I need one more step to get to classification
probabilities.