Fix diversity algorithm and make check happy
The new diversity algorithm multiplies the user-provided weight by the
inverse of the mean number of times each branch leading to the species
has been passed through.

The first species selected is the one with the highest user-provided
score. For each subsequent step, the scores are weighted by diversity
(as described above) and the highest scoring species is selected.

This new algorithm gives nearly identical results to the old one, but
fixes one corner case, where polytomies at equal depth would not be
evenly sampled. E.g.

            0
         /     \
      0          0
    / | \      / | \
   1  1  0    1  0  0
   a  b  c    d  e  f

Assuming the order of leafs in the phylo object is a-f, by the old
algorithm, the remaining leafs would be selected in the order: c,e,f. In
the new algorithm, e,c,f. This makes a big difference in the poorly
resolved NCBI common tree.

One modification I should make in the future, is to weight the diversity
score sum by branch length. This would simply involve a weighted mean
(multiply each count by the normalized branch length).