Replace system2 with system in makeblastdb
When the command given to system2 is not found, system2 croaks without
even returning a value.

For example:

```
> status = 1234
> status = system2("asdfasdf", stderr=TRUE)
> status
1234
```

When `stderr` is not specified, the error message is printed as expected
to the console.

The old `system` command works fine for simple commands, like
`makeblastdb`. It also doesn't actually return the stderr message as a
string, but at least it doesn't pretend to and then not do it.