Export plot_phyloheatmap
Before I was only exporting the overly opinionated "plot_heatmaps"
function which would wrap the heatmaps for all genes in a PDF. The
formatting was dreadful, with many labels appearing behind the heatmap.
PDFs really are not very suitable for these types of graphics. They make
more sense in an HTML format. `plot_phyloheatmap` will directly return
the plots and allow them to be printed in whatever format the use likes.
So you can just dump them all as a list into an RMarkdown file.